fix(Table.stories): correct errors displayed in the browser
